Manual merge VisualisationEngine
